app/api/requirements.txt
Flask>=2.3,<3.0
Flask-SQLAlchemy>=3.0,<4.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2.0




